Remove aria and banner attributes and accessibility fix
Removed the aria label and banner role of the cookie banner div.
Shouldn't really have more than one banner as the <header> already has
the banner role. The aria label is therefore not needed.
Accessibility fix - When clicking on the "I'm OK with cookies" link,
made sure focus next goes on to the banner saved message so screen
readers are given immediate feedback.